Revert "refactor(fsd): Правка структуры и исправление ошибок"
