+++ app/model/appSlice.ts
@@ -3,23 +3,22 @@ import { createSlice } from "@reduxjs/toolkit"
 export const appSlice = createSlice({
   name: "app",
   initialState: {
-    themeMode: "dark", //заглушка
-    error: null as Error,
+    isLoggedIn: false,
   },
   selectors: {
-    selectThemeMode: (state) => state.themeMode,
+    selectIsLoggedIn: (state) => state.isLoggedIn,
   },
   reducers: (create) => ({
-    changeTheme: create.reducer<{ themeMode: string }>((state, action) => {
-      state.themeMode = action.payload.themeMode
+    setIsLoggedIn: create.reducer<{ isLoggedIn: boolean }>((state, action) => {
+      state.isLoggedIn = action.payload.isLoggedIn
     }),
   }),
 })
 
-export const { changeTheme } = appSlice.actions
-export const { selectThemeMode } = appSlice.selectors
+export const { setIsLoggedIn } = appSlice.actions
+export const { selectIsLoggedIn } = appSlice.selectors
 export const appReducer = appSlice.reducer
 
 //types
 export type Error = null | string
-// export type AppInitialState = ReturnType<typeof appSlice.getInitialState>
+export type AppInitialState = ReturnType<typeof appSlice.getInitialState>
