+++ features/auth/assets/img/github-icon.png
