--- shared/lib/random/getRandomNumber.ts
@@ -1,3 +1,0 @@
-export function getRandomNumber(min: number, max: number): number {
-  return Math.floor(Math.random() * (max - min + 1)) + min
-}
