--- shared/ui/index.tsx
@@ -1,2 +1,0 @@
-export { Button } from "./button/button"
-// export { Checkbox } from "./checkbox/checkbox"
